remove unnecessary map_err, include README
